Changed model building
Added required restrictions for appropriate properties.
Added index and alternate key for ApplicationId column of Application.
Renamed Identity tables to be named the same as OpenIddict tables.
Renamed OpenIddictContext to OpenIddictDbContext to emphasis that it is related to Entity Framework as it is done in Identity with IdentityDbContext.

Updated XML documentation to correspond to Identity style.