Updated model building
Overridden names of ASP.NET Identity tables to distinct from other tables and be in the same group as OpenIddict.
Set ClientId as alternate key as it is a unique property that can act as key.
Added name for index on ClientId property.
Set ClientId, RedirectUri and Type properties as required because client application definition cannot be created without this data.